+++ kommons-exec/src/jvmTest/kotlin/com/bkahlert/kommons/exec/CommandLineTest.kt
@@ -2,11 +2,20 @@ package com.bkahlert.kommons.exec
 
 import com.bkahlert.kommons.quoted
 import com.bkahlert.kommons.test.testAll
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.collections.shouldContainExactly
+import io.kotest.matchers.collections.shouldHaveSize
+import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
+import io.kotest.matchers.string.shouldContain
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.io.TempDir
+import java.io.IOException
+import java.nio.file.Path
+import kotlin.io.path.div
+import kotlin.io.path.pathString
 
 class CommandLineTest {
 
@@ -14,6 +23,22 @@ class CommandLineTest {
         CommandLine("command") shouldBe CommandLine(command = "command", arguments = emptyList())
         CommandLine("command", "arg") shouldBe CommandLine(command = "command", arguments = listOf("arg"))
         CommandLine("command", "arg1", "arg2") shouldBe CommandLine(command = "command", arguments = listOf("arg1", "arg2"))
+    }
+
+    @Test fun instantiation(@TempDir tempDir: Path) = testAll {
+        CommandLine(TestClassWithMain::class, "foo", "bar", javaBinary = tempDir / "java", classPath = "foo/bar:baz")
+            .shouldContainExactly(tempDir.resolve("java").pathString, "-cp", "foo/bar:baz", TestClassWithMain::class.qualifiedName, "foo", "bar")
+
+        CommandLine(TestClassWithMain::class, "foo", "bar").exec().readLinesOrThrow() should {
+            it shouldHaveSize 3
+            it[0] shouldBe System.getProperty("java.home")
+            it[1] shouldBe System.getProperty("java.class.path")
+            it[2] shouldBe "foo,bar"
+        }
+
+        shouldThrow<IOException> {
+            CommandLine(TestClassWithoutMain::class, "foo", "bar").exec().readLinesOrThrow()
+        }.message.shouldContain("main method")
     }
 
     @Test fun list() = testAll {
@@ -91,6 +116,20 @@ class CommandLineTest {
     }
 }
 
+
+class TestClassWithMain {
+
+    companion object {
+        @JvmStatic
+        fun main(vararg args: String) {
+            println(System.getProperty("java.home"))
+            println(System.getProperty("java.class.path"))
+            println(args.joinToString(","))
+        }
+    }
+}
+
+class TestClassWithoutMain
 
 internal val CommandLine.Companion.Fixture
     get() = CommandLine("command", "-a", "--bee", "c", " x'\ty'\n💤".quoted)