+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 
 ### Fixed
 
+## [2.3.0] - 2022-10-19
+
+### Added
+
+- Kotest JSON assertions API dependency
+- CommandLine constructor to invoke the main method of class using Java 
+
 ## [2.2.0] - 2022-10-11
 
 ### Added
@@ -52,7 +59,9 @@
 - migrated Kommons Test 0.x.x to this Gradle multi-project
     - [Kommons Test 0.x.x Changelog](https://github.com/bkahlert/kommons-test/compare/v0.1.0...v0.4.4)
 
-[unreleased]: https://github.com/bkahlert/kommons-test/compare/v2.2.0...HEAD
+[unreleased]: https://github.com/bkahlert/kommons-test/compare/v2.3.0...HEAD
+
+[2.3.0]: https://github.com/bkahlert/kommons-test/compare/v2.2.0...v2.3.0
 
 [2.2.0]: https://github.com/bkahlert/kommons-test/compare/v2.1.0...v2.2.0